fix(dcd): refactor and pull useSubmit hook apart
