+++ src/components/[guild]/JoinModal/hooks/useConnectPlatform.ts
@@ -7,7 +7,7 @@ import { env } from "env"
 import { useFetcherWithSign } from "hooks/useFetcherWithSign"
 import usePopupWindow from "hooks/usePopupWindow"
 import useSubmit, { SignedValidation, useSubmitWithSign } from "hooks/useSubmit"
-import { UseSubmitOptions } from "hooks/useSubmit/useSubmit"
+import { UseSubmitOptions } from "hooks/useSubmit/types"
 import { useSetAtom } from "jotai"
 import { OAuthResultParams } from "pages/oauth-result"
 import { useCallback, useMemo } from "react"
