+++ src/components/[guild]/JoinModal/hooks/useMembershipUpdate.tsx
@@ -6,7 +6,7 @@ import useMembership from "components/explorer/hooks/useMembership"
 import useCustomPosthogEvents from "hooks/useCustomPosthogEvents"
 import { useFetcherWithSign } from "hooks/useFetcherWithSign"
 import useSubmit from "hooks/useSubmit"
-import { UseSubmitOptions } from "hooks/useSubmit/useSubmit"
+import { UseSubmitOptions } from "hooks/useSubmit/types"
 import { atom, useAtom } from "jotai"
 import useUsersPoints from "rewards/Points/useUsersPoints"
 import getGuildPlatformsOfRoles from "../utils/getGuildPlatformsOfRoles"
