+++ src/hooks/useSubmit/index.ts
@@ -1,2 +1,2 @@
 export { default, sign, useSubmitWithSign } from "./useSubmit"
-export type { SignedValidation, Validation } from "./useSubmit"
+export type { SignedValidation, Validation } from "./types"
