+++ src/v2/components/Web3ConnectionManager/WalletSelectorModal/hooks/useLinkAddress.ts
@@ -4,7 +4,7 @@ import { AddressLinkParams } from "@/components/Providers/types"
 import { useWeb3ConnectionManager } from "@/components/Web3ConnectionManager/hooks/useWeb3ConnectionManager"
 import { useUserPublic } from "@/hooks/useUserPublic"
 import useSubmit from "hooks/useSubmit"
-import { SignProps } from "hooks/useSubmit/useSubmit"
+import { SignProps } from "hooks/useSubmit/types"
 import { useAtom } from "jotai"
 import randomBytes from "randombytes"
 import { useEffect } from "react"
